Backend/requirements.txt
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Migrate==2.2.1
Flask-SQLAlchemy==2.5
PyJWT==1.4.2
pylint==2.1.1
codecov==2.0.15
coverage==4.5.2
flask-cors==3.0.8
imageio==2.8.0
numpy==1.18.4
pillow==7.1.2
tornado==6.0.4